+++ CHANGELOG.md
@@ -15,5 +15,14 @@
 |1.1.8|2017-02-09|fix: Spinner disappearance stutters|
 |1.1.9|2017-03-14|Sorry I forget what it is about.|
 |1.2.0|2017-03-14|major fix: progress HUD doesn't show when app starts|
+|1.2.1|2017-07-11|fix weak issue|
+|1.2.2|2017-08-07|fix syntax error, may cause syntax errors for your projects|
+|1.2.3|2017-09-07|bqlin: fix progressView can not didmiss bug，fix issuce #8|
+|1.2.4|2017-09-08|bqlin - fix issues #8.|
+|1.2.5|2017-09-10|fix keyboard animation|
+|1.2.6|2017-10-25|fix not dismiss issue|
+|1.2.7|2018-04-10|RealBonus: FTNotification: default dismiss time|
+|1.2.8|2018-04-18|fix: keyboard height issue|
+|1.2.9|2018-05-14|change animations|